implemented network interface

+++ src/me/hakki/nat_project/api/objects/IAgArayuzu.java
@@ -3,5 +3,5 @@ package me.hakki.nat_project.api.objects;
 public interface IAgArayuzu {
     void sogutucuyuAc();
     void sogutucuyuKapat();
-    int sicaklikGonder();
+    float sicaklikGonder();
 }
